Refactor: standardize `@components` imports to lowercase across all files ♻️
- Replaced all `@components` imports to consistently use lowercase paths.
- Updated related stubs, utils, and log messages to align with the new standard.
