+++ resources/js/inertia/components/datatable/TableBody.tsx
@@ -1,6 +1,6 @@
 import { translate } from "@/Models/Translatable";
 import { getNestedPropertyValue } from "@/helper";
-import { TableBodyProps } from "@/components/Datatable/DataTableUtils";
+import { TableBodyProps } from "@/components/datatable/DataTableUtils";
 
 function TableBody<Data>({
   tableSchema,
